float rettet til double

+++ DomainLayer/NewOrderLine.cs
@@ -9,12 +9,12 @@ namespace DomainLayer
 	public class NewOrderLine
 	{
 		private int quantity;
-		private float price;
+		private double price;
 
 		public int Quantity { get { return this.quantity; } set { this.quantity = value; } }
-		public float Price { get { return this.price; } set { this.price = value; } }
+		public double Price { get { return this.price; } set { this.price = value; } }
 
-		public NewOrderLine(int quantity, float price)
+		public NewOrderLine(int quantity, double price)
 		{
 			Quantity = quantity;
 			Price = price;
