+++ DomainLayer/Order.cs
@@ -10,13 +10,13 @@ namespace DomainLayer
 	{
 		private DateTime orderDate;
 		private DateTime deliveryDate;
-		private float totalPrice;
+		private double totalPrice;
 
 		public DateTime OrderDate { get { return this.orderDate; } set { this.orderDate = value; } }
 		public DateTime DeliveryDate { get { return this.deliveryDate; } set { this.deliveryDate = value; } }
-		public float TotalPrice { get { return this.totalPrice; } set { this.totalPrice = value; } }
+		public double TotalPrice { get { return this.totalPrice; } set { this.totalPrice = value; } }
 
-		public Order(DateTime orderDate, DateTime deliverDate, float totalprice)
+		public Order(DateTime orderDate, DateTime deliverDate, double totalprice)
 		{
 			OrderDate = orderDate;
 			DeliveryDate = deliverDate;
